// frontend/src/assets/models/Role.ts
export enum Role {

    UI_DESIGNER = "UI_DESIGNER",
    WEB_DEVELOPER = "WEB_DEVELOPER",
    DISCORD_BOT_DEVELOPER = "DISCORD_BOT_DEVELOPER",
    SOFTWARE_DEVELOPER = "SOFTWARE_DEVELOPER",
    MOBILE_DEVELOPER = "MOBILE_DEVELOPER",
    GAME_DEVELOPER = "GAME_DEVELOPER",
    GRAPHICS_DESIGNER = "GRAPHICS_DESIGNER",
    VFX_DESIGNER = "VFX_DESIGNER",
    ANIMATOR = "ANIMATOR",
    VIDEO_EDITOR = "VIDEO_EDITOR",
    CONTENT_WRITER = "CONTENT_WRITER"

}

export const RoleByOrdinal: Role[] = [
    Role.UI_DESIGNER,
    Role.WEB_DEVELOPER,
    Role.DISCORD_BOT_DEVELOPER,
    Role.SOFTWARE_DEVELOPER,
    Role.MOBILE_DEVELOPER,
    Role.GAME_DEVELOPER,
    Role.GRAPHICS_DESIGNER,
    Role.VFX_DESIGNER,
    Role.ANIMATOR,
    Role.VIDEO_EDITOR,
    Role.CONTENT_WRITER
]